// src/app/application/components/application-form/models/step.model.ts
export enum ApplicationStepRoute {
  step_one = 'application/form/step-one',
  step_two = 'application/form/step-two',
  step_three = 'application/form/step-three',
  step_four = 'application/form/step-four',
  final = 'application/form/final',
}

export enum ApplicationStep {
  step_one = 1,
  step_two = 2,
  step_three = 3,
  step_four = 4,
  final = 5,
}